requirements.txt
# Core dependencies
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.1
pytz==2021.3
requests==2.26.0

# Database
pymongo==4.0.1

# AI and ML
torch==1.10.0
transformers==4.12.5
llama-cpp-python==0.1.65
sentencepiece==0.1.96

# Voice processing
speechrecognition==3.8.1
pyttsx3==2.90
librosa==0.8.1
numpy==1.21.4
scipy==1.7.3

# Health assessment
pandas==1.3.4
scikit-learn==1.0.1

# API and async
gunicorn==20.1.0
asyncio==3.4.3
aiohttp==3.8.1

# Testing
pytest==6.2.5
pytest-cov==2.12.1

# Utilities
tqdm==4.62.3
loguru==0.5.3
pydantic==1.8.2

# Security
pyjwt==2.3.0
cryptography==36.0.0

# Offline sync
sqlite3-to-mysql==1.4.1
apscheduler==3.8.1